backend/requirements.txt
Django
psycopg2
djoser
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
langchain
python-dotenv
